SimpleBLE.jl

Documentation for SimpleBLE.jl

Base.notify — Method

notify(callback,
	peripheral::Peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
)
notify(
	peripheral::Peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
) do data
	# Stuff
end

Set a callback that is called when data is received from a Characteristic. notify is generally faster than indicate because it does not need to acknowledge that it received the data.

Warning

WARNING: YOU MUST NOT PASS THE data VECTOR OUTSIDE THE CALLBACK, IT IS NOT VALID AFTER THE CALLBACK FINISHES, IF YOU NEED TO KEEP THE DATA AFTER THE CALLBACK IS FINISHED, COPY IT!

SimpleBLE.address — Method

address(adapter)

Get the address of an adapter.

See also identifier

SimpleBLE.address — Method

address(peripheral)

Get the address of a peripheral.

See also identifier

SimpleBLE.connect — Method

connect(peripheral) do
	# Stuff
end

Convenience function for automatically disconnecting from a peripheral once you're done with it

SimpleBLE.connect — Method

connect(peripheral)

Connect to a peripheral. Generally you should disconnect when you're done with it

SimpleBLE.find_peripheral — Method

find_peripheral(matchfunc, adapter)
find_peripheral(adapter) do identifier::String
	# Stuff
	return output::Bool
end

Convenience function for finding a peripheral. identifier is the name of the peripheral, it can be an empty string.

SimpleBLE.get_adapter — Method

get_adapter(i)

Acquire a handle to an adapter. i is an index starting at 0. You can check the number of adapters with adapter_get_count.

SimpleBLE.identifier — Method

identifier(adapter)

Get the name of an adapter.

See also address

SimpleBLE.identifier — Method

identifier(peripheral)

Get the name of a peripheral.

See also address

SimpleBLE.indicate — Method

indicate(callback,
	peripheral::Peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
)
indicate(
	peripheral::Peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
) do data
	# Stuff
end

Set a callback that is called when data is received from a Characteristic. notify is generally faster than indicate because it does not need to acknowledge that it received the data.

Warning

WARNING: YOU MUST NOT PASS THE data VECTOR OUTSIDE THE CALLBACK, IT IS NOT VALID AFTER THE CALLBACK FINISHES, IF YOU NEED TO KEEP THE DATA AFTER THE CALLBACK IS FINISHED, COPY IT!

SimpleBLE.manufacturer_data — Method

manufacturer_data(peripheral)

Acquire the manufacturer data associated with peripheral.

SimpleBLE.peripheral_read — Method

peripheral_read(peripheral, service, characteristic)

Read data from a characteristic.

service and characteristic can be AbstractStrings, SBLEUUIDs, or SBLESERVICE and SBLECHARACTERISTIC acquired by services.

SimpleBLE.read_descriptor — Method

read_descriptor(
	peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC},
	descriptor::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
)

Read data from a descriptor.

SimpleBLE.scan_for — Method

scan_for(adapter, timeout_ms)

Scan for a specified number of ms. Results can be acquired with scan_get_results. Or check each peripheral found or updated inside a callback set by set_callback_on_scan_found

See also scan_start, scan_stop

SimpleBLE.scan_get_results — Method

scan_get_results(adapter)

Get the results of a scan.

See also scan_start, scan_stop, scan_for

SimpleBLE.scan_start — Method

scan_start(adapter)

Start scanning for peripherals. Results can be acquired with scan_get_results. Or check each peripheral found or updated inside a callback set by set_callback_on_scan_found

See also scan_stop, scan_for

SimpleBLE.scan_stop — Method

Stop a scan

SimpleBLE.services — Method

services(peripheral)

Acquire the services provided by a peripheral.

SimpleBLE.set_callback_on_connected — Method

set_callback_on_connected(callback, peripheral)
set_callback_on_connected(peripheral) do
	# Stuff
end

Set a callback that is called when a peripheral is connected.

See also set_callback_on_disconnected

SimpleBLE.set_callback_on_disconnected — Method

set_callback_on_disconnected(peripheral) do
	# Stuff
end

Set a callback that is called when a peripheral is disconnected.

See also set_callback_on_connected

SimpleBLE.set_callback_on_power_off — Method

set_callback_on_power_off(adapter) do
	# Stuff
end

Set a callback that is called when an adapter is powered off.

SimpleBLE.set_callback_on_power_on — Method

set_callback_on_power_on(adapter) do
	# Stuff
end

Set a callback that is called when an adapter is powered on.

SimpleBLE.set_callback_on_scan_found — Method

set_callback_on_scan_found(adapter) do peripheral
	# Stuff
end

Set a callback that is called when an adapter finds a peripheral.

See also scan_start, scan_stop, scan_for

SimpleBLE.set_callback_on_scan_start — Method

set_callback_on_scan_start(adapter) do
	# Stuff
end

Set a callback that is called when an adapter starts scanning.

See also scan_start, scan_stop, scan_for

SimpleBLE.set_callback_on_scan_stop — Method

set_callback_on_scan_stop(adapter) do
	# Stuff
end

Set a callback that is called when an adapter stops scanning.

See also scan_start, scan_stop, scan_for

SimpleBLE.set_callback_on_scan_updated — Method

set_callback_on_scan_updated(adapter) do peripheral
	# Stuff
end

Set a callback that is called when an adapter updates a peripheral.

See also scan_start, scan_stop, scan_for

SimpleBLE.unsubscribe — Method

unsubscribe(
	peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC}
)

unsubscribe from a notify or indicate call.

SimpleBLE.write_command — Method

write_command(
	peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC},
	data<: AbstractArray
)

Write a command to a characteristic.

SimpleBLE.write_descriptor — Method

write_descriptor(
	peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC},
	descriptor::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC},
	data <: AbstractArray
)

Write data to a descriptor.

SimpleBLE.write_request — Method

write_request(peripheral,
	service::Union{AbstractString, SBLEUUID, SBLESERVICE},
	characteristic::Union{AbstractString, SBLEUUID, SBLECHARACTERISTIC},
	data<:AbstractArray
)

Write a request to a characteristic.

SimpleBLE.Adapter — Type

A handle for an adapter that enables bluetooth connectivity on the device

Can be acquired with get_adapter(i)

Windows currently only supports one adapter with i = 0

SimpleBLE.Peripheral — Type

A handle for a bluetooth device. You can acquire a specific peripheral by using find_peripheral or get a list of all found peripherals with scan_get_results

SimpleBLE.SBLECHARACTERISTIC — Type

Can be written to with writerequest and writecommand along with it's associated SBLESERVICE

You can check what Characteristics a peripheral has with services Often a peripheral will not advertise it's services unless it is connected

SimpleBLE.SBLEMANUFACTURERDATA — Type

Arbitrary data about a peripheral that can be acquired with manufacturer_data

SimpleBLE.SBLESERVICE — Type

You can check what Services a peripheral has with services. Often a peripheral will not advertise it's services unless it is connected

SimpleBLE.SBLEUUID — Type

Basically just a fixed length string but can be anything really