SimpleBLE.jl

SimpleBLE is a cross-platform bluetooth low energy (BLE) library, designed for simplicity and ease of use.

Simple example

# Get an adapter that will handle requests
adapter = get_adapter(0)
# Find a peripheral using adapter
peripheral = find_peripheral(adapter) do id
    occursin("peripheral name", id)
end
id = identifier(peripheral)
connect(peripheral) do
    SERVICE = "de8d0b82-c7cd-0f33-d15d-c38e1f26673f"
    CHARACTER = "6706b606-0c12-3976-ddba-909377d43dc5"
    notify(peripheral, SERVICE, CHARACTER) do data
        # make sure not to pass the data handle outside the callback
        # it is not valid data once this callback is over, if you need to keep it, copy it
        println("$id received data:")
        println(String(data))
    end
    println("Press enter to stop listening")
    readline()
    unsubscribe(peripheral, SERVICE, CHARACTER)
end

Most common functions

SimpleBLE.get_adapter — Function

get_adapter(i)

Acquire a handle to an adapter. i is an index starting at 0. You can check the number of adapters with adapters_get_count.

SimpleBLE.find_peripheral — Function

find_peripheral(matchfunc, adapter)
find_peripheral(adapter) do identifier::String
	# Stuff
	return output::Bool
end

Convenience function for finding a peripheral. identifier is the name of the peripheral, it can be an empty string.

SimpleBLE.identifier — Function

identifier(adapter)

Get the name of an adapter.

See also address

identifier(peripheral)

Get the name of a peripheral.

See also address

SimpleBLE.address — Function

address(adapter)

Get the address of an adapter.

See also identifier

address(peripheral)

Get the address of a peripheral.

See also identifier

SimpleBLE.connect — Function

connect(peripheral)

Connect to a peripheral. Generally you should disconnect when you're done with it

connect(callback, peripheral)
connect(peripheral) do
	# Stuff
end

Convenience function for automatically disconnecting from a peripheral once you're done with it