Types

SimpleBLE.Adapter — Type

A handle for an adapter that enables bluetooth connectivity on the device

Can be acquired with get_adapter(i)

Windows currently only supports one adapter with i = 0

SimpleBLE.Peripheral — Type

A handle for a bluetooth device. You can acquire a specific peripheral by using find_peripheral or get a list of all found peripherals with scan_get_results

SimpleBLE.SBLECHARACTERISTIC — Type

Can be written to with write_request and write_command along with it's associated SBLESERVICE

You can check what Characteristics a peripheral has with services Often a peripheral will not advertise it's services unless it is connected

SimpleBLE.SBLEMANUFACTURERDATA — Type

Arbitrary data about a peripheral that can be acquired with manufacturer_data

SimpleBLE.SBLESERVICE — Type

You can check what Services a peripheral has with services. Often a peripheral will not advertise it's services unless it is connected

SimpleBLE.SBLEUUID — Type

Basically just a fixed length string but can be anything really